+++ src/server/app.js
@@ -4,7 +4,7 @@ import "regenerator-runtime/runtime";
 import express from "express";
 import bodyParser from "body-parser";
 
-import { route } from "./controllers";
+import { route } from "./routes";
 
 const app = express();
